refactor: extract drawProc into utility
